refactor community actions
